requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
cryptography==36.0.2
Django==4.0.3
django-cors-headers==3.11.0
django-mongodb-engine==0.6.0
djangorestframework==3.13.1
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
idna==2.7
mongoengine==0.24.1
Pillow==9.1.0
pycparser==2.21
pymongo==3.12.3
pyOpenSSL==22.0.0
python-snappy==0.6.1
pytz==2022.1
requests==2.27.1
sqlparse==0.2.4
tzdata==2022.1
urllib3==1.24.3
